vfs: add time from metadata to vfs.Node
Each file has three timestamps: atime, ctime and mtime. These timings
are saved in the metadata. This commit adds new properties and methods
to expose these time stamps in the vfs nodes.

There preexisting attributes [acm]time still exist and have the original
meaning (altough now they are properties). They are the time stamps (in
seconds) as they are determined while reading the git objects.

This commit changes the behaviour of atime, in the sense, that now it
will be set to the same value as [cm]time. Before, this value was always
left at 0.

Having timestamps with seconds precision is a problem, because IEEE
double cannot exactly express the timestamps up to nano seconds
precision. Thus there are the new properties [acm]time_nsec and
[acm]time_nsec_meta. The former is just the [acm]time parameter
in nano seconds, the latter is the timestamp as provided from
the metadata (or None if it doesn't exist).

Then there are new methods [acm]time_nsec_meta_default and
[acm]time_nsec_default which try to return the times and if they are not
available, the return a sensible default.

Signed-off-by: Thomas Haller <thom311@gmail.com>