vfs: add inode to vfs.Node
When metadata is available, a vfs.Node has a hardlink_target. With
this property, we can generate an inode for each node. Of course the
inode cannot be the same as in hardlink_target, because that number
might already be used by another file.

This also makes it possible for nlinks() to return the actual link
count for a node. Note, that this is not really correct until all node
with the same inode are loaded -- because before loading all nodes, we
don't know which nodes will end up having the same hardlink_target.
This means, when you encounter the first node for the same inode, the
inode will be 1. As you encounter more nodes for the same inode, the
nlinks count will increase.

Signed-off-by: Thomas Haller <thom311@gmail.com>